refactor(audit): split red-mode tests into mode_controller_red_test.go
Extract the red-lens test suite (TestReport_RedAnalysisMethods, the
TestRedMode_* battery, and the red-only helpers newRedReport / runRedAnalysis /
findingByTitlePrefix / adminPortalByName) into a dedicated file, cutting
mode_controller_test.go from ~2200 to ~1530 lines. Behavior-preserving: no test
added or removed; blue-mode tests stay in the original file. Addresses the
file-size review finding.
